seperated graphics calls to ant_gl lib

+++ include/util.h
@@ -6,7 +6,7 @@
 ssize_t readLine(int fd, char *buffer);
 void printLine(char *line);
 void swapi(int *x, int *y);
-int max(int x, int y);
-int min(int x, int y);
+int max(float x, float y);
+int min(float x, float y);
 
 #endif
